cbm: unified kernel for all mediatek chromebooks - wip (#17)
this kernel will hopefully replace the mt7 and mt8 kernels soon and
extend them to cover the mt8186 and maybe mt819x chromebooks as well

the kernel config options used were reconstructed from scratch based on
a run of scripts/dtc/dt_to_config on all the mt81xx dts files with a
cleanup and sorting of the options afterwards. that result was compared
to the current mainline defconfig, the defconfig of the collabora
mediatek chromebook ci kernel tree and the old options used for mt7 and
mt8 in the past here. there might be still some fine tuning required,
but it looks quite a bit cleaner than the old config already.

a v6.12.9 kernel built from this was tested to boot fine on my oak-elm,
my kukui-juniper and my corsola-magenton (wifi does not work well yet on
the last one, but that should hopefully be easy to fix soon) ... i did
not do any extensive testing yet, so there might still be some more
regressions included and sound on corsola is still an open topic too ...

no testing has been done on mt819x yet.